requirements.txt
adbutils==0.11.0
allure-pytest==2.9.43
allure-python-commons==2.9.43
apkutils2==1.0.0
Appium-Python-Client==1.2.0
attrs==21.2.0
braga==0.6.2
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
cigam==0.0.3
decorator==5.0.9
Deprecated==1.2.12
deprecation==2.1.0
execnet==1.9.0
facebook-wda==1.4.3
Faker==8.10.1
idna==2.10
iniconfig==1.1.1
logzero==1.7.0
lxml==4.6.3
packaging==20.9
Pillow==8.3.1
pluggy==0.13.1
progress==1.5
py==1.10.0
pyelftools==0.27
pyparsing==2.4.7
pytest==6.2.4
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-random-order==1.0.4
pytest-rerunfailures==10.1
pytest-xdist==2.3.0
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.25.1
retry==0.9.2
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
uiautomator2==2.16.0
urllib3==1.26.6
weditor==0.6.4
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
